Shifted VerifyTokenWithAuthAPI and buildVerifyURL to jwtutils, as it will be resued by other apis too
